refactor(ledger): collocate the asset-rate route registrars :hammer:
Both asset-rate registrars move out of assetrate_handler.go into
assetrate_routes.go, leaving the handler file with the handler struct, the
request/response envelopes and the cores. Asset-rate has no /v2 mirror, so
there are three registrars here rather than four.

No route, authz tuple, operation ID or path changed.
